+++ integration-test/integration/chat/custom_provider_test.clj
@@ -42,7 +42,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "my-provider/foo1"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Tell me a joke!\n"})
@@ -71,7 +71,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "my-provider/foo1"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Who's there?\n"})
@@ -100,7 +100,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "my-provider/foo1"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "What foo?\n"})
@@ -154,7 +154,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "my-provider/deepseekcoder"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Tell me a joke!\n"})
@@ -179,7 +179,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "my-provider/deepseekcoder"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Who's there?\n"})
@@ -204,7 +204,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "my-provider/deepseekcoder"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "What foo?\n"})