+++ integration-test/integration/chat/ollama_test.clj
@@ -34,7 +34,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "ollama/qwen3"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Tell me a joke!\n"})
@@ -60,7 +60,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "ollama/qwen3"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Who's there?\n"})
@@ -80,7 +80,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "ollama/qwen3"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "What foo?\n"})
@@ -125,7 +125,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "ollama/qwen3"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "hello!\n"})
@@ -155,7 +155,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "ollama/qwen3"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "how are you?\n"})
@@ -201,7 +201,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "ollama/qwen3"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "What files you see?\n"})