+++ integration-test/integration/chat/openai_test.clj
@@ -26,7 +26,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "openai/gpt-4.1"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Tell me a joke!\n"})
@@ -55,7 +55,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "openai/gpt-4.1"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "Who's there?\n"})
@@ -84,7 +84,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "openai/gpt-4.1"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "What foo?\n"})
@@ -123,7 +123,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "openai/gpt-5"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "hello!\n"})
@@ -156,7 +156,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "openai/gpt-5"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "how are you?\n"})
@@ -200,7 +200,7 @@
         (is (match?
              {:chatId (m/pred string?)
               :model "openai/gpt-5"
-              :status "success"}
+              :status "prompting"}
              resp))
 
         (match-content chat-id "user" {:type "text" :text "What files you see?\n"})